🔒 Fix Unauthenticated IoT Unlock Endpoint
